codemirror {codemirrorR}	R Documentation

CodeMirror htmlwidget

Description

Create a http://codemirror.netCodeMirror editor for use as an htmlwidgets for rmarkdown, Shiny, the R console, RStudio Viewer,

Usage

codemirror(doc = "", mode = NULL, theme = NULL, ..., width = NULL,
  height = NULL, elementId = NULL)

Arguments

doc
character, file or file path, or xml2 document or nodeset

mode
character specifying a https://codemirror.net/doc/manual.html\#option_modemode

theme
character specifying a https://codemirror.net/doc/manual.html\#option_themetheme

...
arguments for other CodeMirror https://codemirror.net/doc/manual.html\#option_valueoptions

width, height
any valid CSS size unit for the width and height of the container

elementId
character for the unique id of the htmlwidget container; likely NULL unless you need to customize

[Package codemirrorR version 0.0.1 Index]